Full table rewrite based on stockfish TT

+++ search_stack.go
@@ -5,7 +5,7 @@ package zoo
 // The stack is used for reconstructing the searched PV.
 type Stack struct {
 	PV    []Step
-	Depth int16
+	Depth uint8
 	Step  Step
 	Eval  Value
 	Nodes int
